Fixed some compile errors and warnings
<==== Mercurial data ====>
Branch: master

+++ common/src/plugins/resources/onvif/onvif_resource_searcher_wsdd.cpp
@@ -318,7 +318,7 @@ void OnvifResourceSearcherWsdd::findEndpoints(EndpointInfoHash& result)
         }
     }
 
-    bool intfListChanged = intfList.size() != m_ifaceToSock.size();
+    bool intfListChanged = (unsigned int)intfList.size() != m_ifaceToSock.size();
     if( !intfListChanged )
         for( const QnInterfaceAndAddr& intf: intfList )
         {
